fix iframe placeholder html, add test
